Change to real*8

+++ src/bnrydds.f
@@ -47,7 +47,7 @@ c     'x' defined in acom.h
      +             DTGR,TDIS,BCLT      
 	include 'trnsfr.h'
 
-        real*16 frb, tt0, tt, orbits
+        real*8 frb, tt0, tt, orbits
 
         frb = 1.d0/pb(1)
 	an=twopi/pb(1)
